Start on Paxos optimizations.
No longer keep track of old decisions.  We're not applying decisions at
all right now, since we'll delegate that to another actor later on.

+++ src/main/scala/com/comcast/xfinity/sirius/api/impl/paxos/LeaderHelper.scala
@@ -2,7 +2,7 @@ package com.comcast.xfinity.sirius.api.impl.paxos
 import com.comcast.xfinity.sirius.api.impl.paxos.PaxosMessages._
 
 class LeaderHelper {
-  def proposalExistsForSlot(proposals: Set[Slot], newSlotNum: Int) = proposals.exists {
+  def proposalExistsForSlot(proposals: Set[Slot], newSlotNum: Long) = proposals.exists {
     case Slot(existingSlotNum, _) if existingSlotNum == newSlotNum => true
     case _ => false
   }
@@ -11,13 +11,13 @@ class LeaderHelper {
   def update[T](x: Set[T], y: Set[T]) = y ++ x.filterNot(y.contains(_))
 
   def pmax(pvals: Set[PValue]): Set[Slot] = {
-    def updateWithMaxBallot(acc: Map[Int, PValue], pval: PValue) = acc.get(pval.slotNum) match {
+    def updateWithMaxBallot(acc: Map[Long, PValue], pval: PValue) = acc.get(pval.slotNum) match {
       case None  => acc + (pval.slotNum -> pval)
       case Some(PValue(otherBallot, _, _)) if pval.ballot >= otherBallot => acc + (pval.slotNum -> pval)
       case _ => acc
     }
 
-    val maxBallotMap = pvals.foldLeft(Map[Int, PValue]())(updateWithMaxBallot)
+    val maxBallotMap = pvals.foldLeft(Map[Long, PValue]())(updateWithMaxBallot)
     maxBallotMap.foldLeft(Set[Slot]()) {
       case (acc, (_, PValue(_, slotNum, proposal))) => acc + Slot(slotNum, proposal)
     }
